+++ src/System.Windows.Forms.Primitives/src/Interop/Richedit/Interop.ENPROTECTED.cs
@@ -1,0 +1,21 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal static partial class Richedit
+    {
+        [StructLayout(LayoutKind.Sequential, Pack = 4)]
+        public struct ENPROTECTED
+        {
+            public User32.NMHDR nmhdr;
+            public int msg;
+            public nuint wParam;
+            public nint lParam;
+            public CHARRANGE chrg;
+        }
+    }
+}
